requirements.txt
flask
flasgger
python-dotenv
pymongo